==> 0_obtain_data.ipynb <==
import dataset_tools as dtools

# Import HAM10000 dataset
# Source: https://www.kaggle.com/datasets/kmader/skin-cancer-m
dtools.download(dataset="Skin Cancer: HAM10000", dst_dir="~/dataset/HAM10000/")

import kagglehub
from kagglehub import KaggleDatasetAdapter

# Set the path to the file you'd like to load
file_path = "/dataset/ISIC2020/file.csv"

# Load the latest version
df = kagglehub.load_dataset(
    KaggleDatasetAdapter.PANDAS,
    "sumaiyabinteshahid/isic-challenge-dataset-2020",
    file_path,
    # Provide any additional arguments like
    # sql_query or pandas_kwargs. See the
    # documenation for more information:
    # https://github.com/Kaggle/kagglehub/blob/main/README.md#kaggledatasetadapterpandas
)

print("First 5 records:", df.head())
